fix integer error

+++ crates/scouter_sql/src/sql/schema.rs
@@ -354,7 +354,7 @@ pub struct LLMDriftServerSQLRecord {
 
     pub processing_ended_at: Option<DateTime<Utc>>,
 
-    pub processing_duration: Option<i64>, // Interval in seconds for the drift calculation
+    pub processing_duration: Option<i32>,
 }
 
 impl LLMDriftServerSQLRecord {
